use more defaults for language configuration
Previously we set hide_untranslated=False, which is not the default in django-cms and the django-cms default makes more sense.

With this change the default override is removed and all other
settings which matched the defaults anyway were removed.

django-parler adopts the settings from django-cms anyway, if
PARLER_LANGUAGES is not defined, so we let parler do it's thing as well
by removing our settings.